Add missing mock util fix for test path; add windows specific mock test
